requirements.txt
fastapi
uvicorn
pyserial
pyserial-asyncio
aiofiles
websockets
